+++ Source/Base/ASAvailability.h
@@ -56,11 +56,8 @@
   #define YOGA __has_include(YOGA_HEADER_PATH)
 #endif
 
-// When enabled, use ASTextNode2 for ALL instances of ASTextNode.
-// This includes what ASButtonNode uses internally, as well as all app references to ASTextNode.
-// See ASTextNode+Beta.h declaration of ASTextNodeExperimentOptions for more details.
-#ifndef ASTEXTNODE_EXPERIMENT_GLOBAL_ENABLE
-  #define ASTEXTNODE_EXPERIMENT_GLOBAL_ENABLE 0
+#ifdef ASTEXTNODE_EXPERIMENT_GLOBAL_ENABLE
+  #error "ASTEXTNODE_EXPERIMENT_GLOBAL_ENABLE is unavailable. See ASConfiguration.h."
 #endif
 
 #define AS_PIN_REMOTE_IMAGE __has_include(<PINRemoteImage/PINRemoteImage.h>)