+++ Source/Details/ASGraphicsContext.h
@@ -32,15 +32,6 @@ NS_ASSUME_NONNULL_BEGIN
 ASDISPLAYNODE_EXTERN_C_BEGIN
 
 /**
- * Call this to enable the experimental no-copy rendering.
- *
- * Returns YES if it was enabled, or NO + assert if it's too late because
- * rendering has already started. In practice it's fine to call this
- * during -didFinishLaunchingWithOptions:.
- */
-extern BOOL ASEnableNoCopyRendering(void);
-
-/**
  * Creates a one-shot context.
  *
  * Behavior is the same as UIGraphicsBeginImageContextWithOptions.